requirements.txt
absl-py==2.1.0
altair==5.3.0
antlr4-python3-runtime==4.9.3
anyio
appnope
APScheduler==3.10.4
argon2-cffi
argon2-cffi-bindings
asttokens
astunparse==1.6.3
async-lru
attrs
Babel
beautifulsoup4
bleach
blinker==1.8.2
Bottleneck
Brotli
cachetools==5.3.3
certifi
cffi
charset-normalizer
click==8.1.7
comm
contourpy
cssselect2==0.7.0
cvzone==1.6.1
cycler
debugpy
decorator
deepface==0.0.92
defusedxml
dlib==19.24.4
executing
face-recognition==1.3.0
face_recognition_models==0.3.0
fastjsonschema
filelock
fire==0.6.0
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
flatbuffers==24.3.25
fonttools
fpdf==1.7.2
fsspec==2024.6.0
gast==0.5.4
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2
google-pasta==0.2.0
grpcio==1.64.1
gunicorn==22.0.0
h5py==3.11.0
html5lib==1.1
idna
imageio==2.34.2
imageio-ffmpeg==0.5.1
imbalanced-learn==0.12.3
ipykernel
ipython
ipywidgets
itsdangerous==2.2.0
jedi
Jinja2
joblib
json5
jsonschema
jsonschema-specifications
jupyter
jupyter-console
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab-pygments
jupyterlab-widgets
jupyterlab_server
keras==3.3.3
kiwisolver
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe
matplotlib
matplotlib-inline
mdurl==0.1.2
mistune
ml-dtypes==0.3.2
mpmath
mtcnn==0.1.1
namex==0.0.8
nbclient
nbconvert
nbformat
nest-asyncio
networkx==3.3
notebook
notebook_shim
numexpr
numpy
omegaconf==2.3.0
opencv-contrib-python==4.10.0.82
opencv-python==4.10.0.82
opt-einsum==3.3.0
optree==0.11.0
overrides
packaging
pandas
pandocfilters
parso
pexpect
pillow
platformdirs
plotly
ply==3.11
prometheus-client
prompt-toolkit
protobuf==4.25.3
psutil
ptyprocess
pure-eval
py-cpuinfo==9.0.0
pyarrow==16.1.0
pycparser
pydeck==0.9.1
pydyf==0.11.0
pygame==2.6.0
Pygments
pyparsing
pyphen==0.15.0
PyQt5==5.15.10
PyQt5-sip
PySocks
python-dateutil
python-json-logger
pytz
PyYAML
pyzmq
qtconsole
QtPy
referencing
requests
retina-face==0.0.17
rfc3339-validator
rfc3986-validator
rich==13.7.1
rpds-py
scikit-learn==1.5.1
scipy

